requirements.txt
dotenv
asgiref==3.3.1
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
decorator==4.4.2
defusedxml==0.7.0rc1
Django==3.1.4
django-debug-toolbar==3.1.1
django-debug-toolbar-force==0.1.8
django-filter==2.4.0
django-nine==0.2.3
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-yasg==1.20.0
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
ipython==7.19.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.8
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
pycparser==2.20
Pygments==2.7.3
PyJWT==2.0.0
pyparsing==2.4.7
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.2
wcwidth==0.2.5
zipp==3.4.0
